requirements.txt
tkinter                  # For GUI components (standard with Python)
pillow                   # For handling images
sqlite3                  # For database (standard with Python)
python-dotenv             #for storing environment variables